+++ source/frontends/sa2/gamepad.h
@@ -13,7 +13,7 @@ public:
   Gamepad(const int index);
 
   virtual bool getButton(int i) const;
-  virtual int getAxis(int i) const;
+  virtual double getAxis(int i) const;
 
 private:
   std::shared_ptr<SDL_GameController> myController;
